jira-agent/requirements.txt
google-adk==1.5.0
mcp==1.10.1
llama-index==0.12.43
